Allow GitHub API token not to be passed
On some CI machines, there's a default `GITHUB_TOKEN` that is left
unset. Prior to this change, we'd see that there's that environment
variable, and attempt to initialize our GitHub API client with it. This
breaks every API call however, because the GitHub API client decides to
send an empty string as an API token on every call in this particular
case.

Now, if that environment variable exists but is unset, we don't try to
send an empty string as an API token.